SWEP breaking fix.
as outlined in Issue #193.
First of all, i've changed the StripWeapon, so it won't be removing the
fresh bought weapon. Afterwards, this run into some problems with
GetNWInt, which used the new weapon instead of the old weapon.
This was fixed by adding an new variable attached to the plymeta, in
which the old weapon is stored as soon an buy is happening. On L357 of
sv_weps is also an failsafe, incase the function get called without
using one of the other methods, to still function properly.