Napalm Zombie Boss Extension
Exactly what it says on the tin, an optional extension which puts the
napalm zombie into nZombies.

Napalm will spawn in on round 7 and every 2-3 rounds after that, he will
walk at a slow pace towards the nearest player and will explode if
killed or if a player is too close.
If a player kills him, that player will get 750 points.

In BO1 he begins sprinting if left alive after a round ends, but that
isn't the case here... yet.