Black Ops 3 Styled HUD
So I was reading through one of the projects and saw an idea from Dj_Nathan that said "perks near rounds like BO3?"
I commented on that (albeit nobody replied since it was like a month old or something) and posted example screenshots.

This here is the actual code I changed for my personal use.

CHANGES:
Round tallies only go up to five, changing to a number at 6 and beyond. This is done because the tally marks would overlap the perk icons, looking bad.

Perks are shown near the round, just like in BO3 and WW2.

NOTES:
If you buy too many perks, (like 11 or more) then the icons will overlap with the text for power-ups. Shouldn't be a problem with the default perk limit of 4, and can go up to 10 without visual annoyance.